rename more conflicting files
