+++ sdk/formrecognizer/azure-ai-formrecognizer/samples/v3.1/async_samples/sample_convert_to_and_from_dict_v3_1_async.py
@@ -7,15 +7,15 @@
 # --------------------------------------------------------------------------
 
 """
-FILE: sample_convert_to_and_from_dict.py
+FILE: sample_convert_to_and_from_dict_v3_1_async.py
 
 DESCRIPTION:
-    This sample demonstrates how to convert models returned from an analyze operation
+    This sample demonstrates how to convert models returned from a recognize operation
     to and from a dictionary. The dictionary in this sample is then converted to a
     JSON file, then the same dictionary is converted back to its original model.
 
 USAGE:
-    python sample_convert_to_and_from_dict.py
+    python sample_convert_to_and_from_dict_v3_1_async.py
 
     Set the environment variables with your own values before running the sample:
     1) AZURE_FORM_RECOGNIZER_ENDPOINT - the endpoint to your Form Recognizer resource.
@@ -24,11 +24,13 @@ USAGE:
 
 import os
 import json
+import asyncio
 
-def convert_to_and_from_dict():
+async def convert_to_and_from_dict_async():
     path_to_sample_forms = os.path.abspath(
         os.path.join(
             os.path.abspath(__file__),
+            "..",
             "..",
             "..",
             "./sample_forms/id_documents/license.jpg",
@@ -37,7 +39,8 @@ def convert_to_and_from_dict():
 
     from azure.core.serialization import AzureJSONEncoder
     from azure.core.credentials import AzureKeyCredential
-    from azure.ai.formrecognizer import FormRecognizerClient, RecognizedForm
+    from azure.ai.formrecognizer.aio import FormRecognizerClient
+    from azure.ai.formrecognizer import RecognizedForm
 
     endpoint = os.environ["AZURE_FORM_RECOGNIZER_ENDPOINT"]
     key = os.environ["AZURE_FORM_RECOGNIZER_KEY"]
@@ -45,10 +48,11 @@ def convert_to_and_from_dict():
     form_recognizer_client = FormRecognizerClient(
         endpoint=endpoint, credential=AzureKeyCredential(key)
     )
-    with open(path_to_sample_forms, "rb") as f:
-        poller = form_recognizer_client.begin_recognize_identity_documents(identity_document=f)
-    
-    id_documents = poller.result()
+    async with form_recognizer_client:
+        with open(path_to_sample_forms, "rb") as f:
+            poller = await form_recognizer_client.begin_recognize_identity_documents(identity_document=f)
+        
+        id_documents = await poller.result()
 
     # convert the received model to a dictionary
     recognized_form_dict = [doc.to_dict() for doc in id_documents]
@@ -78,5 +82,9 @@ def convert_to_and_from_dict():
     print("----------------------------------------")
 
 
-if __name__ == "__main__":
-    convert_to_and_from_dict()
+async def main():
+    await convert_to_and_from_dict_async()
+
+
+if __name__ == '__main__':
+    asyncio.run(main())
